pyproject.toml
[project]
name = "purple-merit-assessments"
version = "0.1.0"
description = "Purple Merit AI/ML Engineer assessments — launch war room + bug triage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["assessment2_bug_triage/mini_repo"]
